chore: reinstating tubular to prod-ready for retirement
these user retirement scripts haven't been updated since the re-factor
back into  edx-platform a year ago. This is catching up with all of the
interim changes. It's not just a mass copy, for two reasons: the file
paths and package paths for imports are different; and this is
configured as a solo package with a setup.py.

the changes include:

* the updates in removing _some_ ancient functionality and workarounds specific to old versions of Python.
* Switching to the  included version of `mock`, per above
* Removing all code related to the Demographics pipeline
* updating or removing pip constraints
* ran `make upgrade`  to get newer pip files.
* a re-factor that was apparently necessary for `moto` updates. I believe this is a  package that allows mocking the AWS `boto` library.
* updating the package to run Python 3.11 on ubuntu-latest  (in github workflows, in `setup.py`, in the `Dockerfile`, etc.)

This does not include any formatting updates, and it does not add any changes that were not already added to edx-platform. So, for example, there are some ancient uses of `from six import ...`  that have not been removed. There are also some unordered imports, and other ancient functionality. The goal here is to just get this matching the version on platform as much as possible.

FIXES:
* master